refactor clan domain
